Adds support for a `debug` flag in version.py.  This can be set by passing the `--debug` option to `./setup.py build` (note: this is not supported by `./setup.py install` install, so you must manually run `build` to create a debug build).  If `./setup.py build` is run without `--debug` then the debug flag is set to `False`.  If the debug value changes from the previous build, a rebuild of all C extensions is forced.
There are at least two advantages to having `debug` in version.py.  The first
is informational: Whether or not the package was built with debug enabled is
frozen with the source code.  The other advantage is that the flag becomes
globally available to modules that need to configure C extensions, and it could
even be used by pure Python code, for example, to enable debug logging
globally.

This patch went through several iterations before I was happy with it.  The
Cython changes are mostly incidental at this point, though in an earlier
version of the patch they were more necessary.  This is because the `build` and
`build_ext` commands were being subclassed.

However, I eventually realized that in order for that approach to work
consistently it would also be necessary to subclass all of the other `build_`
commands.  And the more you start overriding things in distutils the further
you start going down the `numpy.distutils` path, which is a direction best
avoided.

This version of the patch still uses the distutils machinery to get the value
of the --debug option.  This has the advantage that it can also take the debug
option from one's setup.cfg file.  For example, adding

 [build]
 debug = True

to setup.cfg during development ensures that debug mode is always turned on,
for example.