Add Row.__array__ method for np.array(row) conversion
Used similar test as for Table.__array__, except that for the
Row.__array__ method it was necessary to return np.array(self._data)
instead of just self._data.  Otherwise self._data is an np.void and
this makes numpy complain that what came back from __array__ is not
an array.  So np.array(row) never returns a ref to the original data,
as reflected in the test_convert_numpy_array() function.

As an unexpected bonus, the previously failing test_right_equal()
test started working.  Go figure.