Remove wcsconfig.h from version control.
Because this file can be updated depending on the platform being built on,
it shouldn't be under version control (it gets annoying because git picks
up on any modifications made to it).  Since it will be generated at build
time if it doesn't already exist, then even more reason for it to not be
there to begin with.