This lessens the need for `__init__.py` files to check `is_in_build_mode()` to optionally turn off importing other things in their package that may require 2to3 or C extensions to have been built.
The real cause of `__init__.py` files being imported at build time was due to a `setup_package.py` file within that package being imported.  By importing the file using `imp.load_module()` rather than `importlib.import_module()`, we can import the file as a standalone module without causing an import of its parent package as a side effect.  This allows (in many cases) for the `__init__.py` to not need to block the import of 2to3 or C extension code because the `__init__.py` is not longer imported at build time.

However, there is still a special case in `astropy.wcs` where the package itself is needed at build time, so a call to `is_in_build_mode()` has been added there.  However, that case is an exception, not the norm.

Also in this change: The setup script currently tries to import the legacy module (vo, pyfits, pywcs) to see if it's in fact `pyfits` or astropy's legacy shim that masquerades as `pyfits`.  However, if it's the latter, and it's imported at build time, and `astropy` has been uninstalled but its `pyfits` legacy shim has not been (a corner case if there ever was one), importing the legacy shim will result in importing pyfits from the source tree which doesn't work because it requires 2to3 and has C extensions.  The change here is to rather than import the legacy shim, to just open it as a file and look for the magic variable name `_is_astropy_legacy_alias`.