Fix handling of missing / incorrect scale arg
Previously the error reporting was wrong when the scale arg was not
supplied for cases where the format was auto-determined.  This commit
fixes that and also cleans up the error messages.  It also adds
tests of this functionality along with a new test that simply creates
a time object in all currently available formats.