Change table tests to use absolute imports
Currently the table tests use pytest fixtures which were introduced
in pytest 2.3.  The astropy version is 2.2.4.  On at least my mac
installation I was unable to use an external pytest 2.3 with test
scripts that reside within the astropy directory because pytest
ended up loading the py package from astropy (not my system version)
and that version is not recent enough.  The workaround to get
testing to work is installing astropy to a virtualenv and running
the tests within a directory completely outside the astropy source
tree.  This requires absoute imports.