Support units and format in io.ascii (initial use-case of Daophot)
This commit modifies the TableOutputter to copy any io.ascii column format and
units attributes to the corresponding attrs in the output Table columns.  The
initial use case is Daophot, which includes both units and format info in the
default specification.  Some side effects of this change cropped up and are
dealt with here.

Reading the format and units info for Daophot required modifying the
ContinuationLinesInputter to have a no_continue attribute that provides
a regex which disables line continuation when the regex matches.  It
also required rewriting the header parsing for Daophot.

The Daophot table is also the basis for all write tests, which is good because
adding the format info highlighted a known issue that when writing tables with
format specifiers like "%10.2f", the extra spaces generated by that output
fooled the Python CSV writer into thinking the spaces had to be quoted.  This
is a little tricky to deal with because in theory a table could have values
with significant whitespace on the ends which cannot be stripped.  This forced
adding a new strip_whitespace kwarg to ui.write and iu.get_writer that allows
explicit control in case the default behavior is not correct.

The latex splitter join method had to be modified to strip extra whitespace
from the formatting.

A number of test write outputs changed slightly because of the application
of the proper value formatting.

Conflicts:

	astropy/io/ascii/core.py