Improve reading of DAOPhot tables according to discussion in #510
This includes a fix to enable include_names and exclude_names for
Daophot tables and adds tests for this capability.
Also, some tests failed in the previous version, because the code now
requires strict adherence to the DAOPHOT output format, while
previously the testfile daophot.dat was missing a few white spaces.
Thus, this test data has been fixed.