Update CHANGES.rst
Added an empty changelog section for 0.2 line of release.  Any changes that will be backported to 0.2.x that have been made since 0.2b1 should be listed there.  I need to go through myself and make sure this is properly groomed before doing the next 0.2.x release.  Note that assuming the next such release is 0.2b2 then everything under this 0.2 section will become the changelog for 0.2b2, and a new 0.2 section will be added.  At some point this process needs to be better documented.