Check Column/MaskedColumn __new__ args for compliance with 0.3 ordering
This is implemented as a decorate on __new__(cls, name=None, data=None, ..)
for Column and MaskedColumn.

The warning is:

In the next major release of astropy (0.3), the order of function
arguments for creating a {class_name} will change.  Currently the order is
{class_name}(name, data, ...), but in 0.3 and later it will be
{class_name}(data, name, ...).  This is consistent with Table and NumPy.

In order to use the same code for Astropy 0.2 and 0.3, column objects
should be created using named keyword arguments for data and name, e.g.:
{class_name}(name='a', data=[1, 2]).