Fix tests so no warnings are generated in Column() calls
This is only complete for table and io.ascii.  All of these tests
run without generating warnings.  This was done to be sure that
all code in astropy.table and astropy.io.ascii is fully compliant
with the new Column args checking, i.e. it will never happen that
a call to astropy.Table can generate a warning to the user.