Change join tests to compare sorted table outputs
The row order in the join output is driven by details in the
numpy sort routine.  This ends up being different between Mac
and Linux.  To handle this the table outputs are sorted as a
list before being compared.  Since the join row order is not
guaranteed to be consistent across platforms this is OK.