address #910 (fill_values in astropy.io.ascii.write)
This implementation solves the problem discussed in #910.
I suspect that it costs some performance for large tables (I did
not actually perform tests for that), because I replaced an
iterator with an explicet list.
However, in this way most of the fill_values logic can be reused
from the implemention of fill_values on input.