fix bug in writing AASTex tables
The old implementation did not honor `latexdict['tabletype']` for AASTex tables
due to the order in which the various defaults were set.
This fix restores that and makes it possible to write e.g. 2-column
deluxetable* tables.
Small changes in existing tests are sufficient to make sure this bug will not
creep in again.