Fixes a bug in get_format and adds has_format
get_format could fail in weird ways because it would check if 'get_format' was a valid format.
This fix simply ignores the last two entries in the ``__all__`` list ('get_format', and the just added 'has_format')
to prevent this issue. An alternative solution could be to use introspection.

has_format returns a boolean as to whether the passed in format would work or not.