Tick angles point along gridlines regardless of aspect ratio
I had previously computed the tick angle by advancing a fixed distance
in axes coordinates. This was wrong, because the axes coordinates always
run from 0 to 1, regardless of aspect ratio. The solution in this patch
is to compute the offsets by advancing in figure coordinates (i.e.,
pixels).

Fixes astrofrog/wcsaxes#62.