Slight clarification of this 'hack' I think (and avoids creating one unnecessary list).
I was thinking of adding this recipe [1] to Astropy's utils and using it
to compute the MRO, which I think is a bit nicer.  But some performance
testing showed it not to be significantly faster in most cases, so I
stuck with the temp class creation which is not too horrible after all.

[1] http://code.activestate.com/recipes/577748-calculate-the-mro-of-a-class/