add a test to see if pixels in the projected plane are non-square (distorted)
Add a function that tests a WCS to see if pixels, when projected to the
projection plane, remain square or if they become non-square
(distorted).

Simply checking for the presence of distortion properties (`sip`,
`det2im1`, etc.) is not enough to conclude whether or not pixels will
look square in the projection plane - see #3112. In addition, having
equal pixel scales (see, e.g., #3230) also does not guarantee that
pixels will be square.

This function considers that pixels will be "square" on the projection
plane if:

1. all distortion properties of the WCS object (`sip`, `det2im1`, etc.)
are `None`; and

2. The CD matrix corresponding to the celestial axes is orthogonal (or
unitary for complex-valued CD matrices - not supported at present).

If one of the above conditions is not satisfied then pixels will be
non-square.