Remove _col_update_attrs_from and remove explicit info.copy()
Previously the idiom to copy info from one object the next was:

  obj.info = other.info.copy()

This had a problem because it was just setting obj.info to
the info from other, but this only works if obj and info have
the same Info class.  So instead the info setter is basically
just the same as the copy method.