Rebased the table indices work from #3915 on top of my Column.__getitem__ speedup branch of #4075, and made adjustments to take advantage of it in the indices code.  This included getting rid of the `def get_item(...)` hacks, which I think are no longer necessary--the base Column.__getitem__ is now almost exactly as fast as the stock implementation in the basic case of single item scalar access.
In the process I also refactored the Cython code a bit so that it could
be less repetitive.

For now, tests that failed due to the change in behavior (indicies now
*are* copied by default when slicing) are commented out; but all other
tests pass.

It should also that the 'copy_on_getitem' index mode is now a no-op
since that's the default behavior.  We could still change this so that
it's not the default behavior, if we prefer, or add a no_copy_on_getitem
mode instead.