More improvements (?) to the _ASTROPY_SETUP_ detection and auto-build mechanism.
As of Astropy helpers v1.1, the astropy_helpers module itself sets up
the _ASTROPY_SETUP_ global where appropriate (see
https://github.com/astropy/astropy-helpers/pull/191/files).  This means
it's no longer necessary to have the _ASTROPY_SETUP_ boilerplate in
setup.py.

In general this shouldn't fail.  But sometimes when accidentally
building astropy with an old astropy-helpers, this can blow up in
spectacular ways.  It's conceivable this could fail in other ways too,
but I haven't though of them yet.

This patch ensures that the _ASTROPY_SETUP_ global is always set
correctly, even if it somehow was not set by astropy_helpers (by
using the same mechanism that astropy_helpers uses to determine
the flag).