Should fix the reference leak between FITS_rec and Column.array as explained here: https://github.com/astropy/astropy/pull/520/files#diff-8ba48829876ba755e67054a1a90ec124R2367
I will add more in-line documentation of this in a subsequent commit.
But basically what's going on is this:  Whenever a Column object's
.array object is assigned to, it actually walks that array's bases to
check for the reference cycle, by seeing if one of those bases is a
FITS_rec instance with the present Column in its ColDefs.  If so, rather
than creating a hard reference it stores a weak reference to the
original FITS_rec, and the .array property returns the array via that
weak reference.

That way the reference cycle never gets created in the first place.

On the FITS_rec end, it actually also keeps a set of weak references to
all Column objects associated with that FITS_rec through its ._coldefs.
When FITS_rec.__del__ is called (because its reference count has gone to
zero) it deletes self._coldefs.  If that ColDefs object has no
references outside the FITS_rec (i.e. a user has not pointed a variable
to it), it and all Columns it contains are properly deleted.  This also
removes the column from the FITS_rec's _col_weakrefs set.

However, if there are still references to the Column (again, by a user
for example) then the data for that array is copied (maybe instead we
can just restore the reference to the original FITS_rec?), so the user
can continue to access that data.  This is allowed because by the time
FITS_rec.__del__ is called its underlying memory has not been freed yet.
The one exception to this was where data.deleter closed any mmap
underlying the data.  But I fixed the order of things so that the mmap
is still available to perform that copy.

This is all probably still fairly fragile, but in the vast majority of
cases it fixes the chance for any reference leaks.