Disable initial ordering in jsviewer.
By default dataTables sort in ascending order with the first column, though it
seems more appropriate to keep the original table order.

Ref http://datatables.net/reference/option/order

Also ref #4404 which added an index column. When this column is present the
original order is respected, but then a useless sort is done by dataTables so it
is also a good idea to disable the initial sort.