Fix sigma_clipped_stats to use the axis argument
The axis argument was added in #3792 and is passed correctly to sigma_clip, but
then mean, median and std did not use axis and were computed on the whole array.
This commit changes this to use axis to compute the statistics.

Also another minor annoyance is that, with Numpy 1.10, np.ma.median always
return a MaskedArray, even when axis is None. This was not the case in previous
version (and will probably be fixed in the future), and mean and std return
a scalar. So if it happens here, we take the first element of the array.