Update equivalencies.py
Multiply instead of divide in the equation for beam area (beam_area = 2.*np.pi*(bmaj*bmin*fwhm_to_sigma**2)), because of the way fwhm_to_sigma is defined with (8*np.log(2))**0.5 already in the denominator.  

I think this is consistent with the equations shown on this page: https://www.cv.nrao.edu/course/astr534/Interferometers2.html  But, you can ignore this if I have made a mistake or wrong assumption.