Handle backwards compatibility for current behavior of accessing HDUs after a file has been closed.
Now you can open a file, close it, and still read the headers (but not the data---that had been disabled for a long time) after it was closed.  This of course means that before the file is closed it reads in any HDUs that haven't already been read, killing much of the benefit gained from lazy-loading in the first place.

However this is *not* performed in certain special cases where a user would never manually (either by calling close() or using a with statement) close the file--namely the convenience functions.

However, use of the old behavior does raise a deprecation warning.  This should give users the chance to fix poorly written code that 'closes' and HDUList before it's done being used.  Then we can take fuller advantage of the lazy loading.

It might also be worth adding an option to disable the backwards-compat mode so that one can take full advantage of lazy-loading immediately.