Fix FITS ASCII to float conversion for columns with E/F types
ASCII columns with the E and F data type (TFORM) are mapped to `float32`
by default (with the `ASCII2NUMPY` dict), but can be mapped to `float64`
if necessary. This is done in `_convert_ascii_format`, and the resulting
Numpy format is stored as `.recformat` on the `_AsciiColumnFormat`
object.

But this `recformat` attribute was not used when the real conversion
happen, later in `FITS_rec._convert_ascii`. The default `float32` format
was used instead.