Prevent PLY to overwrite its generated files.
Ref #5581, fixes #5726.

After #5526 updated PLY to 3.9, some files were re-generated at runtime
(astropy/coordinates/{angle_lextab.py,angle_parsetab.py}). This is because of
this line https://github.com/dabeaz/ply/blob/master/ply/lex.py#L901 which
assumes that the lextab parameter in `lex.lex` is a str only, not unicode, which
leads to some surprise when unicode_literals is used on Python 2 !

So this commit just wraps the lextab value to make sure that a str is always
passed to `lex.lex`.