Improve speed of coordinate transformations for arrays of obstime
The erfa.astrom struct needs a substantial amount of computing power
to update Earth parameters. However, all but the Earth rotation angle
are relatively slowly changing numbers. If some accuracy can be
sacrificed, calculating the slowly changing astrom parameters
only on certain time-steps, can lead to large speed-ups. For example,
a time resolution of 10 minutes will still have an accuracy of <~50 mas
if ICRS is transformed to AltAz.

add astrom_manager helpers and transform_precision context

coordinates.astrom_manager: do interpolation of earth_pv and CIS

just for curiosity, we removed the pure MJD-binning in get_astrom helper
and implemented a linear interpolation of the Earth-PV, Heliocentric, and
CIS (Cartesian) vectors based on a binned MJD support. Comparing to the
MJD binning, the accuracy of the coordinate transforms is boosted by a
factor of a thousand, or so, while the computing times are almost as
fast as before.