Represent a few more numpy types in YAML
This addresses Issue #6076.
There is some problem with pyyaml versions. The one found on Github and Bitbucket has a `represent_long` function, but the one published on pypi doesn't, even if it should be more recent. So, depending on the version that the user has installed, they will use the proper `represent_long` or revert to `represent_int`, in the latter case with a warning.