Add unit tests, attempt to fix FITS file mode handling...
This attempts to implement a more sensible (and thorough) approach to
hanhandling of FITS file modes. The docstring for fitsopen already says
that it is not valid to pass an open file handle with a mode that is not
compatible with the given FITS file mode, but this was not being tested.
This also required a fix to a different unit test that was breaking the
documented rule--this error went unchecked until now.