Fix testing of rst files in cases where temporary directory path includes a component with a leading underscore.
The test collection for the rst files formerly got rid of any rst file that included a part of the path with a leading underscore. This was originally intended to filter out files in e.g. a _build directory. However, this also got rid e.g. of /tmp/_q/docs/index.rst. This commit fixes this by only doing this filtering for folders with a leading underscore inside the docs/ directory. Likewise, filtering done based on the 'api' folder name is only done inside the docs/ directory.