Import all asdf tags in extension.py, not with __init__.py...
This is to make sure that all tag implementations are imported by the
time we create the Astropy extension class so that the extension is
correctly populated with the provided types. The reason for not doing
this in __init__.py is to avoid pytest collection errors in the case
that asdf is not installed.