Change preferred CDS formatting of arcsec, arcmin.
Corrects a failed test after changes in PR #7348. The CDS
specification accepts both 'arcs' and 'arcsec' for arcseconds (and
similarly for arcminutes), see
http://vizier.u-strasbg.fr/cgi-bin/Unit. 'arcsec' seems to be the
preferred form (http://cds.u-strasbg.fr/doc/catstd-3.2.htx, but
an argument could be made the other way via 'rule 2' on that page)
so make that the default string representation.