use a hash to decide when to use the cached representation_info
This is necessary because at any time the user might add new representation or differential classes.  Other pacakges (e.g. Sunpy) use this functionality in the wild so it must be supported.  This change uses a hash of the rep/diff class dictionaries to detect when a change has occurred.