Simplify Quantity._set_own_unit.
As it was, it counted too much on the input already being a Quantity
or normal float/array. Hence, for a class that implemented to_value
but did not have a normal dtype it would fail.

Also add a few more general tests for Quantity initialization from
a class that has a unit attribute.