Validate jd1, jd2 and value.
jd1 and jd2 must be arrays of floats, and if one is an array they
will be broadcast together so that they are the same shape. If any of
this is not true an exception will be raised.

value can be more or less anything and if jd1/jd2 are zero-dimensional
then value will be unboxed and converted to a conventional python type
with a few exceptions. Masking prevents unboxing, and neither datetime64
nor record dtypes are converted to python equivalents.

Finally the otherwise hard-to-debug error of failing to mark value as a
property is caught at the time of construction of a new custom format
class.