Ensure at least one test trips the "W48 Unknown attribute" warning.
Restores one line of lost test coverage by inserting a bad <OPTION>
attribute into the gemini.xml test.  Previously none of the tests
actually had this particular XML fault, but the code bug fixed in
io/votable/tree.py on the present PR was erroneously calling the
error handler for it; now we have a test that really earned it.