Update docstring & keywords for isclose and allclose
I corrected the one-liner at the top of the docstring for
astropy.units.isclose, since it was the one-liner that was meant for
np.allclose.  While I was at it, I expanded the docstrings to list the
parameters, describe the requirements on the units of `rtol` and
`atol`, and add a section on what exceptions are raised.

I made the `equal_nan` keyword argument explicit since it has been in
numpy for several years now (I think).  Since `isclose` and `allclose`
both took `**kwargs` already, I don't think it changes anything.

I set the default value of `atol` to zero instead of `None` in order
to be more explicit about what the defaults are.  I updated the error
messages to f-strings because they are awesome.