Document clusters, remote_timeout=0
Users can now ask that astropy not attempt to download files from the
internet by setting an option in their configuration files.

There is now documentation on how to manage file downloading on
clusters.

Now the remote-data=none setting will squawk if we try to touch the
internet with remote_timeout=0.