Make frame attribute equivalency checks symmetric
The `BaseCoordinateFrame._frameattr_equiv(left_fattr, right_fattr)`
check is not symmetric with respect to the two attributes.

If the left attribute is a coordinate frame or representation and the
right one isn't, the method returns False. If the right attribute is a
coordinate frame or representation and the left one isn't, one is at the
mercy of the underlying equality operator instead. This typically
results in a TypeError (or a FutureWarning in the case of an
EarthLocationAttribute vs an unspecified None).

This was encountered when calculating the angular separation between
a star and the Sun in a common AltAz frame. The Sun SkyCoord had an
extra `obsgeoloc` frame attribute due to its GCRS origins while the
star SkyCoord had this as None. Then `Sun.separation(star)` works but
`star.separation(Sun)` raises a TypeError.