[#5408] Add regression test and fix for #5408
I had to rework how Card._verify works: Normally the run_option method,
when passed a fix= argument, runs the fix code immediately.  However,
the fix code can have side-effects that modify the behavior of the
later verifications, which makes reasoning about _verify very difficult.

So I changed it so that it first collects a list of errors (and
potentially errors to fix) and then runs the fixes collectively at the
end.  It would probably be better if that were the way verify() worked
in general, but that is a larger refactoring job to leave for later
(possibly can wait until #3668).