Fix #11298
The original code made the mistake of assuming that the number of
groups the string will be split into is directly related to the length
of the input or the number of spaces, but this is false if the input
is made up of multiple long words that need to be split across lines as
in the regression test.

In this case, if a long word follows another long word we want to keep
the existing semantics of breaking after a space, and then beginning
the next word on a new line.