Fix test failing after backport of #11285
With the backport of #11285 on v4.2, a new warning is rasied and turned
into an error. This doesn't happen on v4.0 because the way to catch
warning changed (using pytest) and does not happen on master/4.3 because
some new code was added to detect if the file is a valid FITS file.