Update VOTable documentation [ci skip]
The VOTable documentation now contains references to the Unified File
Read/Write Interface, which are analogous to the references already
present in the FITS documentation.

Furthermore, the documentation currently claims that Astropy supports
VOTable versions up to 1.3 proposal, even though support for version 1.4
has been implemented (issue #9474 and pull request #9475). This commit
updates the VOTable version support information. It is possible that the
documentation contains more outdated information and it might benefit
from being reviewed by someone that is familiar with the VOTable
implementation in Astropy.