more improvements to the introduction and a bit of extended technical detail
(I did not see a mention of the term "prefix" anywhere which is at least
good for *NIX users to be exposed to)

I admit I fudged the details on Windows a bit, but if anyone wants to
expand on it, in short virtualenv on Windows essentially copies some of
the contents of C:\PythonX.Y (if that's where you installed it
originally), and IIRC also makes some changes by way of
site_customize.py.  But I have not used virtualenv on Windows in a long
time so I'm hazy on the precise details.